project_comments/requirements.txt
asgiref==3.8.1
bleach==6.0.0
Django==5.1.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
Faker==13.3.4
pillow==11.0.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
webencodings==0.5.1
